+++ includes/cli/class-initializer.php
@@ -1,0 +1,42 @@
+<?php
+/**
+ * Newspack plugin CLI initializer
+ *
+ * @package Newspack
+ */
+
+namespace Newspack\CLI;
+
+use WP_CLI;
+
+defined( 'ABSPATH' ) || exit;
+
+/**
+ * Initializer CLI commands
+ */
+class Initializer {
+
+	/**
+	 * Initialized this class and adds hooks to register CLI commands
+	 *
+	 * @return void
+	 */
+	public static function init() {
+		add_action( 'init', [ __CLASS__, 'register_comands' ] );
+	}
+
+	/**
+	 * Adds CLI commands. Do not call directly or before init hooks
+	 *
+	 * @return void
+	 */
+	public static function register_comands() {
+		if ( ! defined( 'WP_CLI' ) ) {
+			return;
+		}
+
+		WP_CLI::add_command( 'newspack setup', 'Newspack\CLI\Setup' );
+
+	}
+
+}