+++ newspack.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name: Newspack
  * Description: An advanced open-source publishing and revenue-generating platform for news organizations.
- * Version: 1.95.4
+ * Version: 1.99.0
  * Author: Automattic
  * Author URI: https://newspack.com/
  * License: GPL2
@@ -14,7 +14,7 @@
 
 defined( 'ABSPATH' ) || exit;
 
-define( 'NEWSPACK_PLUGIN_VERSION', '1.95.4' );
+define( 'NEWSPACK_PLUGIN_VERSION', '1.99.0' );
 
 // Load language files.
 load_plugin_textdomain( 'newspack-plugin', false, dirname( plugin_basename( __FILE__ ) ) . '/languages' );
@@ -23,6 +23,8 @@ load_plugin_textdomain( 'newspack-plugin', false, dirname( plugin_basename( __FI
 if ( ! defined( 'NEWSPACK_PLUGIN_FILE' ) ) {
 	define( 'NEWSPACK_PLUGIN_FILE', __FILE__ );
 }
+
+require_once 'vendor/autoload.php';
 
 // Include the main Newspack class.
 if ( ! class_exists( 'Newspack' ) ) {